Rename, refactor and remove 4D
